backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "aiofiles>=25.1.0",
    "aiosqlite>=0.22.1",
    "alembic>=1.18.4",
    "celery>=5.6.3",
    "fastapi>=0.136.0",
    "google-genai>=1.73.1",
    "httpx>=0.28.1",
    "passlib[bcrypt]>=1.7.4",
    "pillow>=12.2.0",
    "playwright>=1.58.0",
    "pydantic>=2.13.2",
    "pydantic-settings>=2.14.0",
    "pyjwt>=2.12.1",
    "python-multipart>=0.0.26",
    "python-slugify>=8.0.4",
    "redis>=7.4.0",
    "sqlalchemy>=2.0.49",
    "uvicorn[standard]>=0.44.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
pythonpath = ["."]
